+++ forges/josh-gerrit-changes/src/lib.rs
@@ -1,0 +1,7 @@
+//! Gerrit publishing support for Josh
+
+mod publish;
+mod refs;
+
+pub use publish::*;
+pub use refs::*;
