+++ forges/josh-gerrit-changes/src/refs.rs
@@ -1,0 +1,37 @@
+//! Parsing of Gerrit-style push refnames for josh's git transports: splitting
+//! `%` push options and translating the Gerrit magic refs (`refs/for`,
+//! `refs/drafts`, `refs/publish/for`) onto real branches, deciding which josh
+//! `PushMode` a push triggers.
+
+use anyhow::anyhow;
+
+use josh_changes::PushMode;
+
+pub fn baseref_and_options(
+    refname: &str,
+    author: &str,
+) -> anyhow::Result<(String, String, Vec<String>, PushMode)> {
+    let mut split = refname.splitn(2, '%');
+    let push_to = split.next().ok_or(anyhow!("no next"))?.to_owned();
+
+    let options = if let Some(options) = split.next() {
+        options.split(',').map(|x| x.to_string()).collect()
+    } else {
+        vec![]
+    };
+
+    let mut baseref = push_to.to_owned();
+    let mut push_mode = PushMode::Normal;
+
+    if baseref.starts_with("refs/for") {
+        baseref = baseref.replacen("refs/for", "refs/heads", 1)
+    }
+    if baseref.starts_with("refs/drafts") {
+        baseref = baseref.replacen("refs/drafts", "refs/heads", 1)
+    }
+    if baseref.starts_with("refs/publish/for") {
+        push_mode = PushMode::Publish(author.to_string());
+        baseref = baseref.replacen("refs/publish/for", "refs/heads", 1)
+    }
+    Ok((baseref, push_to, options, push_mode))
+}
