--- josh-changes/src/forges/mod.rs
@@ -1,3 +1,0 @@
-pub mod gerrit;
-
-pub use gerrit::*;
