+++ josh-changes/src/lib.rs
@@ -1,6 +1,6 @@
 //! Stacked-changes metadata store for josh: change identity, comments, votes,
-//! revisions, and stacked-changes push machinery (including Gerrit-style
-//! publishing), persisted in `refs/josh/...` refs.
+//! revisions, and stacked-changes push machinery, persisted in `refs/josh/...`
+//! refs.
 //!
 //! The public API is flat: everything is re-exported at the crate root.
 
@@ -8,7 +8,6 @@ pub use josh_core::trailers::{commit_change_meta, parse_change_meta};
 
 mod change;
 mod comments;
-mod forges;
 pub mod layout;
 mod refs;
 mod revisions;
@@ -18,7 +17,6 @@ mod votes;
 
 pub use change::*;
 pub use comments::*;
-pub use forges::*;
 pub use refs::*;
 pub use revisions::*;
 pub use stacked::*;
