+++ josh-cli/Cargo.toml
@@ -40,6 +40,7 @@ josh-github-auth.workspace = true
 josh-github-keyring.workspace = true
 josh-github-graphql.workspace = true
 josh-github-changes.workspace = true
+josh-gerrit-changes.workspace = true
 josh-graphql.workspace = true
 josh-link.workspace = true
 josh-templates.workspace = true
