+++ josh-proxy/Cargo.toml
@@ -56,6 +56,7 @@ gix-hash.workspace = true
 url.workspace = true
 
 josh-changes.workspace = true
+josh-gerrit-changes.workspace = true
 josh-rpc.workspace = true
 josh-core.workspace = true
 josh-templates.workspace = true
